'(', ')' 처리문제

+++ app/src/main/java/com/example/mobile_programming_term_project/makeCalculationStack.java
@@ -3,7 +3,7 @@ package com.example.mobile_programming_term_project;
 public interface makeCalculationStack {
     boolean isEmpty();
     boolean isFull();
-    void push(int item);
-    int pop();
-    int peek();
+    void push(float item);
+    float pop();
+    float peek();
 }
